# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fathom",
    "version": "1.0.0",
    "description": "Macroeconomic data API powered by FRED (Federal Reserve Economic Data). 816,000+ time series covering GDP, inflation, employment, interest rates, money supply, and more.",
    "x-guidance": "Use /api/popular to discover available indicators. Use /api/search to find specific series by keyword. Use /api/observations to fetch data values. Use /api/snapshot for a dashboard overview. Use /api/compare to compare multiple series. Series IDs are uppercase codes like GDP, UNRATE, CPIAUCSL, FEDFUNDS, DGS10.",
    "contact": {
      "email": "admin@orbonomy.xyz"
    },
    "x-logo": {
      "url": "https://fathom.vercel.app/favicon.png"
    }
  },
  "servers": [
    {
      "url": "https://fathom.vercel.app"
    }
  ],
  "paths": {
    "/api/observations": {
      "post": {
        "operationId": "observations",
        "summary": "Get economic data observations",
        "description": "Fetch data values for any FRED economic series. Use series IDs like GDP, UNRATE, CPIAUCSL, FEDFUNDS, DGS10.",
        "tags": [
          "Data"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.030000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "seriesId": {
                    "type": "string",
                    "minLength": 1,
                    "description": "FRED series ID (e.g. GDP, UNRATE, CPIAUCSL)"
                  },
                  "limit": {
                    "type": "string",
                    "description": "Max observations (default: 10)"
                  },
                  "sort": {
                    "type": "string",
                    "description": "Sort order: asc or desc (default: desc)"
                  },
                  "start": {
                    "type": "string",
                    "description": "Start date YYYY-MM-DD"
                  },
                  "end": {
                    "type": "string",
                    "description": "End date YYYY-MM-DD"
                  },
                  "units": {
                    "type": "string",
                    "description": "Units transformation (lin, chg, ch1, pch, pca, cch, cca, log)"
                  },
                  "frequency": {
                    "type": "string",
                    "description": "Frequency (d, w, bw, m, q, sa, a)"
                  }
                },
                "required": [
                  "seriesId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Get economic data observations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "series_id": {
                      "type": "string"
                    },
                    "count": {
                      "type": "number"
                    },
                    "observations": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/categories": {
      "post": {
        "operationId": "categories",
        "summary": "Browse data categories",
        "description": "Browse FRED data categories: National Accounts, Prices, Employment, Money & Banking, etc.",
        "tags": [
          "Discovery"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.010000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categoryId": {
                    "type": "string",
                    "description": "Category ID (default: 0 for top-level)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Browse data categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "category_id": {
                      "type": "string"
                    },
                    "categories": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/category-series": {
      "post": {
        "operationId": "categoryseries",
        "summary": "Series in a category",
        "description": "List economic data series within a specific FRED category, sorted by popularity.",
        "tags": [
          "Discovery"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.020000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categoryId": {
                    "type": "string",
                    "minLength": 1,
                    "description": "FRED category ID"
                  },
                  "limit": {
                    "type": "string",
                    "description": "Max results (default: 20)"
                  }
                },
                "required": [
                  "categoryId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Series in a category",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "category_id": {
                      "type": "string"
                    },
                    "series": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/bulk": {
      "post": {
        "operationId": "bulk",
        "summary": "Multi-series fetch",
        "description": "Fetch observations for multiple FRED series in one call. Returns latest values for each.",
        "tags": [
          "Data"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.050000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "seriesIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Array of FRED series IDs (e.g. ['GDP', 'UNRATE', 'CPIAUCSL'])"
                  },
                  "limit": {
                    "type": "string",
                    "description": "Observations per series (default: 5)"
                  }
                },
                "required": [
                  "seriesIds"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Multi-series fetch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "results": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/calendar": {
      "post": {
        "operationId": "calendar",
        "summary": "Economic calendar",
        "description": "Upcoming economic data release calendar with dates and release names.",
        "tags": [
          "Discovery"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.030000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "limit": {
                    "type": "string",
                    "description": "Max results (default: 15)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Economic calendar",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/category-tags": {
      "post": {
        "operationId": "categorytags",
        "summary": "Tags for a category",
        "description": "Get all tags associated with a FRED data category.",
        "tags": [
          "Metadata"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.010000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categoryId": {
                    "type": "string",
                    "minLength": 1,
                    "description": "FRED category ID"
                  }
                },
                "required": [
                  "categoryId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tags for a category",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "category_id": {
                      "type": "string"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/category-tree": {
      "post": {
        "operationId": "categorytree",
        "summary": "Category tree",
        "description": "Get child categories for a parent FRED category. Walk the category tree.",
        "tags": [
          "Discovery"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.010000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categoryId": {
                    "type": "string",
                    "description": "Parent category ID (default: 0 for root)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category tree",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "parent_id": {
                      "type": "string"
                    },
                    "children": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    }
  }
}